[#26] refactor : register api v1 분리
